Created a framework for event broadcasting across processes. The parent process has a message broker, which is simply a zmq sub socket that forwards everything it gets to a zmq pub socket. The details of this server are propagated to child processes, which may instantiate Event objects. Event objects have a post() and a wait() method. Events are named, and when post() is called, any Events (possibly in different processes) that are wait()ing on the same name will return. Otherwise the wait() method blocks. There is also an 'id' argument to the wait() and post() methods, which allows you to ensure that things are in sync.
For example, in BLACS, we will have a few events posted per shot. If something
crashes, we don't want processes to respond to events that were left over
from the last shot. So we will use the hdf5 filename as a unique identifier, and
when we wait on an event, those with an unmatching filename will be ignored.