Zlock bugfix: lock objects were in wrong state after failed acquisition.
The exception handing code for a failed acquisition was supposed to decrement
a counter (self.recursion_level -= 1) to keep track of the recursion level
so the locks could be used re-entrantly. Instead, it was setting the the counter
to negative one (self.recursion_level =- 1), leading to incorrect behaviour
after a failed lock acquisition.