Removed functionality for running in main thread.
ZMQ recv() doesn't appear to be interruptable with ctrl-c in windows.
This is unacceptable, so we will always run the server in a thread so we can
interrupt it from the main thread by calling context.term().

To prevent the need for boilerplate to run forever and catch a keyboard
interrupt in the main thread, we now provide shutdown_on_interrupt()
which will block until the user hits ctrl-c and then shutdown the server cleanly.

Bumped version to 1.3.3