Do not leak ZPROCESS_PARENTINFO to subprocesses, and remove them from the current environment so that they are not visible in process explorer etc.
This prevents leaking of the shared secret to subprocesses, and prevents the
shared secret being read casually in process_explorer, though it is still there
for a brief time at startup. However, one needs permission to read the process's
memory in order to obtain its environment, so this limits those who can read the
environment to the user who owne the process, or the administrator, so is not a
security issue. However, not leaking the environment to subprocesses is
importatnt, since they could be started as another user that does not have
permission to read its parent's environment.