Addition of Interruptor class, for interrupting blockings sends and recvs. Make ReadQueue and WriteQueue use one for blocking operations, and have startup code pass a specific interruptor instance in that will be used by terminate() to interrupt startup code if terminate() is called during startup.
Have the parent pass the startup timeout to the child process so that the child
process can use the startup timeout to communicate with the parent. This
way if terminate() is called and the parent abandons the child during a remote
Popen, when the child doesn't exist yet, the child will time out and
exit with an exception rather than living on as a zombie waiting forever.