More kosher `start_daemon()` function.
call os.waitpid() on unix to wait for the spawned process to exit.
use CREATE_NO_WINDOW on windows to ensure that the child process
actually has a console, but that it is hidden. This ensures any
subprocesses it itself creates will inherit the hidden console and not
create any more. Use os.DEVNULL for all handles on all platforms.

Remove broken embed() function, yet to be replaced.