Enable IPv6 by default again.
With fixes for issues previously encountered when I last tried this.

Ensure that when we look up a peer's IP address, if the peer is talking
to us over IPv4, that we use its IPv4 address instead of a IPv4-mapped
IPv6 address.

Ensure that when we split a tcp endpoint into host and port numbers,
that we split only on the last colon, to allow for IPv6 addresses
(which contain colons).