Minor changes to task repetition in Task and TaskQueue
Tasks have an `__eq__` method that compares an `_id` attribute, this is
preserved across `copy` which is done when a repeated task is re-added
to the queue.

This can be used to check if a task is in the queue with e.g. `if task
in tasks`, which will evaluate to `True` if a copy of the task is in
the queue, as it would be for a repeated task.

Also, manually re-adding a task to the queue resets its `_called`
attribute so it doesn't raise an exception when next called.

These changes are to support holding a reference to a task that has
`repeat=True`, and being able to check if (a copy of) that task is in
the queue, and being able to cancel and re-add it.