Fix memory leak in Interruptor.unsubscribe(), affecting ZMQClient.get()
The unsubscribe request was not actually processed, leading it to fill a
queue, leaking memory over time (looks like ~64 bytes per call on my
system).

Add a `self._xpub.recv()` call to ensure the unsubscribe request is
processed.

This necessitates acquiring the instance's lock, since `self._xpub` is
not thread-local.